Extracting byte range parsing
