+++ fs/range_test.go
@@ -1,0 +1,9 @@
+package fs_test
+
+import (
+	. "github.com/onsi/ginkgo"
+)
+
+var _ = Describe("ParseByteRanges", func() {
+	XIt("parses byte ranges in other formats -- drill down into a lower level piece")
+})
